refactor: stats mutation factory
